backend/requirements.txt
fastapi==0.135.2
uvicorn[standard]==0.42.0
python-multipart==0.0.20
reportlab>=4.2,<5
oracledb>=2.1.2,<3
pydantic-settings>=2.2.0,<3
email-validator>=2.0.0,<3
python-jose[cryptography]>=3.3.0,<4
passlib[bcrypt]>=1.7.4,<2
